event handler updates
Factor a common proc from the event setup proc that takes a switch for whether ot not to set the max block time.  From the event setup function, invoke it with the time-setting enabled.  From the event check function, invoke it with the time-setting disabled.

This seems to stop the runaways because the check function can respond to changes in zookeeper interest.  But it needs to not call Tcl_SetMaxBlockTime because it causes it to spin a fair amount (not a complete runaway) when the countdown timer gets close to 0.